Restructure tidysic package
